creating lightin prefab and light objects
Created a light mask prefab that can be dragged onto any object to put a halo of light around it.  Also created our basic prototype light objects that have this light as a prefab abject. Scattered them around the map, and stored them all in a "light objects" object for organization